fix(runtime): remove redundant u32 cast on launch_spec.id
CI clippy 1.91 flags `launch_spec.id as u32` as unnecessary since
the field is already u32. Drop the cast.

+++ crates/gglib-runtime/src/process/manager.rs
@@ -424,7 +424,7 @@ impl ProcessManager {
                             if let Err(e) = std::fs::create_dir_all(slot_dir) {
                                 tracing::warn!("Failed to create slot directory: {}", e);
                             }
-                            purge_stale_slot_bin_files(slot_dir, launch_spec.id as u32);
+                            purge_stale_slot_bin_files(slot_dir, launch_spec.id);
                         }
 
                         let port = {
